parse metaphor epl leaf
